refactor: extract init related functions to separate file
