+++ i18nilize/src/internationalize/command_line.py
@@ -1,5 +1,4 @@
 import argparse
-import os
 
 from src.internationalize import globals
 from src.internationalize.helpers import (
@@ -7,11 +6,11 @@ from src.internationalize.helpers import (
     add_update_translated_word,
     delete_translation,
 )
-from src.internationalize.package_setup import (
-    create_directories,
-    setup_directories_exist,
+from src.internationalize.package_init_utils import (
+    initialize_root_directory,
+    setup_package,
+    validate_required_directories,
 )
-from src.internationalize.project_root_utils import get_project_root_directory
 from src.internationalize.sync_processor import pull_translations, push_translations
 
 
@@ -62,7 +61,7 @@ def cli():
         setup_package()
         return
 
-    validate_setup_directories(globals.ROOT_DIRECTORY)
+    validate_required_directories()
 
     if args.command == "add-language":
         add_language(args.language)
@@ -78,32 +77,6 @@ def cli():
         push_translations()
     else:
         print("Invalid command.")
-
-
-def initialize_root_directory():
-    try:
-        root_directory = get_project_root_directory()
-        globals.ROOT_DIRECTORY = root_directory
-        globals.LANGUAGES_DIR = os.path.join(root_directory, "languages")
-    except FileNotFoundError as err:
-        print("Error:", err)
-        exit(1)
-
-
-def validate_setup_directories(root_directory):
-    if not setup_directories_exist(root_directory):
-        print(
-            'Error: i18nilize has not been setup yet. Run "i18nilize setup" to initialize the package.'
-        )
-        exit(1)
-
-
-def setup_package():
-    try:
-        create_directories()
-    except FileExistsError as err:
-        print("Error:", err)
-        exit(1)
 
 
 if __name__ == "__main__":
